separate memberController
