+++ src/main/java/com/Moleugo/moleugo/controller/member/LoginController.java
@@ -1,4 +1,4 @@
-package com.Moleugo.moleugo.controller;
+package com.Moleugo.moleugo.controller.member;
 
 import com.Moleugo.moleugo.entity.Member;
 import com.Moleugo.moleugo.response.LoginResponse;
@@ -6,39 +6,20 @@ import com.Moleugo.moleugo.service.MemberService;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.ApplicationContext;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.CookieValue;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 
-@Slf4j
 @Controller
 @RequiredArgsConstructor
-public class MemberController {
+public class LoginController {
     private final MemberService memberService;
     private final ApplicationContext ac;
-
-    @PostMapping("/signup/check-form")
-    public ResponseEntity<?> isValidForm(@RequestBody Member member) {
-        return ResponseEntity.status(memberService.isValidForm(member)).build();
-    }
-
-    @PostMapping("/signup/verify-email")
-    public void sendVerificationEmail(@RequestBody Member member) {
-        memberService.sendVerificationEmail(member);
-    }
-
-    @GetMapping("/signup/{uuid}")
-    public String signUp(@PathVariable("uuid") String uuid) {
-        if(memberService.signUp(uuid) == HttpStatus.OK) {
-            return "redirect:/#/signup-success";
-        }
-        else {
-            return "redirect:/#";
-        }
-    }
 
     @PostMapping("/login")
     public ResponseEntity<?> login(@RequestBody Member member, HttpServletResponse resp) {
